Move event validation from lambda to endpoint.
